runtimes/nutera/requirements.txt
# Core dependencies
PyYAML==6.0.3

# LLM providers
openai==2.28.0
google-generativeai>=0.3.0

# Optional: for DeepSeek (uses OpenAI-compatible API)
# openai is already included above

# Development and testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
